add note section
