+++ .idea/misc.xml
@@ -118,8 +118,10 @@
         <entry key="common/uicomponent/src/main/java/app/tinks/readkeeper/uicomponent/cellview/ReadingProgressCircle.kt" value="0.64375" />
         <entry key="common/uicomponent/src/main/java/app/tinks/readkeeper/uicomponent/cellview/TimeText.kt" value="0.5070323488045007" />
         <entry key="common/uicomponent/src/main/java/app/tinks/readkeeper/uicomponent/detail/ActionSection.kt" value="0.335" />
+        <entry key="common/uicomponent/src/main/java/app/tinks/readkeeper/uicomponent/detail/InfoSection.kt" value="0.38333333333333336" />
         <entry key="common/uicomponent/src/main/java/app/tinks/readkeeper/uicomponent/detail/ReadingVipInfoSection.kt" value="0.3175925925925926" />
         <entry key="common/uicomponent/src/main/java/app/tinks/readkeeper/uicomponent/detail/ReadingVipNoteSection.kt" value="0.20081135902636918" />
+        <entry key="common/uicomponent/src/main/java/app/tinks/readkeeper/uicomponent/detail/ReadingVipProgressSection.kt" value="0.337037037037037" />
         <entry key="common/uicomponent/src/main/java/app/tinks/readkeeper/uicomponent/detail/actionsection/ReadingActionSection.kt" value="0.45" />
         <entry key="common/uicomponent/src/main/java/app/tinks/readkeeper/uicomponent/detail/actionsection/SearchActionSection.kt" value="0.3990740740740741" />
         <entry key="common/uicomponent/src/main/java/app/tinks/readkeeper/uicomponent/list/ListBookInfo.kt" value="0.25" />
