+++ common/uicomponent/src/main/java/app/tinks/readkeeper/uicomponent/Section.kt
@@ -11,7 +11,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.sp
 
-
 @ExperimentalMaterial3Api
 @Composable
 fun Section(
